Revert "Merge pull request #6752 from andyque/refactorLayout"
This reverts commit ae118a9d22cd7ec0e2656ef605361467fe0f64cf, reversing
changes made to 6f2f2046c7729b48d69a3c2f54c94fefc4145c8a.
