+++ cocos/ui/UILayout.h
@@ -30,27 +30,14 @@ THE SOFTWARE.
 NS_CC_BEGIN
 
 namespace ui {
-    
-class LayoutExecutant;
-    
-class LayoutProtocol
-{
-public:
-    LayoutProtocol(){}
-    virtual ~LayoutProtocol(){}
-    
-    virtual LayoutExecutant* createLayoutExecutant() = 0;
-    virtual const Size getLayoutContentSize() = 0;
-    virtual const Vector<Node*>& getLayoutElements() = 0;
-    virtual void doLayout() = 0;
-};
 
 /**
  *  @js NA
  *  @lua NA
  */
+class LayoutExecutant;
     
-class Layout : public Widget, public LayoutProtocol
+class Layout : public Widget
 {
     
     DECLARE_CLASS_GUI_INFO
@@ -326,11 +313,7 @@ protected:
     
     void setStencilClippingSize(const Size& size);
     const Rect& getClippingRect();
-    
-    virtual void doLayout()override;
-    virtual LayoutExecutant* createLayoutExecutant()override;
-    virtual const Size getLayoutContentSize()override;
-    virtual const Vector<Node*>& getLayoutElements() override;
+    virtual void doLayout();
     
     //clipping
     void onBeforeVisitStencil();
@@ -342,11 +325,12 @@ protected:
     void updateBackGroundImageColor();
     void updateBackGroundImageOpacity();
     void updateBackGroundImageRGBA();
+    LayoutExecutant* createCurrentLayoutExecutant();
     
     /**
      *get the content size of the layout, it will accumulate all its children's content size
      */
-    Size getLayoutAccumulatedSize() const;
+    Size getLayoutContentSize() const;
     
     /**
      * When the layout get focused, it the layout pass the focus to its child, it will use this method to determine which child 
@@ -492,6 +476,8 @@ protected:
     
     Color3B _backGroundImageColor;
     GLubyte _backGroundImageOpacity;
+    
+    LayoutExecutant* _curLayoutExecutant;
     
     GLint _mask_layer_le;
     GroupCommand _groupCommand;
