+++ cocos/ui/UIScrollView.cpp
@@ -49,7 +49,7 @@ ScrollInnerContainer* ScrollInnerContainer::create()
     CC_SAFE_DELETE(widget);
     return nullptr;
 }
-
+    
 const Size& ScrollInnerContainer::getLayoutSize()
 {
     Widget* parent = getWidgetParent();
@@ -210,7 +210,7 @@ void ScrollView::setInnerContainerSize(const Size &size)
         }
         case Direction::HORIZONTAL:
         {
-            if (_innerContainer->getRightBoundary() <= _size.width)
+            if (_innerContainer->getRightInParent() <= _size.width)
             {
                 Size newInnerSize = _innerContainer->getSize();
                 float offset = originalInnerSize.width - newInnerSize.width;
@@ -223,7 +223,7 @@ void ScrollView::setInnerContainerSize(const Size &size)
             Size newInnerSize = _innerContainer->getSize();
             float offsetY = originalInnerSize.height - newInnerSize.height;
             float offsetX = 0.0f;
-            if (_innerContainer->getRightBoundary() <= _size.width)
+            if (_innerContainer->getRightInParent() <= _size.width)
             {
                 offsetX = originalInnerSize.width - newInnerSize.width;
             }
@@ -233,11 +233,11 @@ void ScrollView::setInnerContainerSize(const Size &size)
         default:
             break;
     }
-    if (_innerContainer->getLeftBoundary() > 0.0f)
+    if (_innerContainer->getLeftInParent() > 0.0f)
     {
         _innerContainer->setPosition(Vec2(_innerContainer->getAnchorPoint().x * _innerContainer->getSize().width, _innerContainer->getPosition().y));
     }
-    if (_innerContainer->getRightBoundary() < _size.width)
+    if (_innerContainer->getRightInParent() < _size.width)
     {
          _innerContainer->setPosition(Vec2(_size.width - ((1.0f - _innerContainer->getAnchorPoint().x) * _innerContainer->getSize().width), _innerContainer->getPosition().y));
     }
@@ -245,7 +245,7 @@ void ScrollView::setInnerContainerSize(const Size &size)
     {
         _innerContainer->setPosition(Vec2(_innerContainer->getPosition().x, _innerContainer->getAnchorPoint().y * _innerContainer->getSize().height));
     }
-    if (_innerContainer->getTopBoundary() < _size.height)
+    if (_innerContainer->getTopInParent() < _size.height)
     {
         _innerContainer->setPosition(Vec2(_innerContainer->getPosition().x, _size.height - (1.0f - _innerContainer->getAnchorPoint().y) * _innerContainer->getSize().height));
     }
@@ -390,56 +390,56 @@ bool ScrollView::checkNeedBounce()
     {
         if (_topBounceNeeded && _leftBounceNeeded)
         {
-            Vec2 scrollVector = Vec2(0.0f, _size.height) - Vec2(_innerContainer->getLeftBoundary(), _innerContainer->getTopBoundary());
+            Vec2 scrollVector = Vec2(0.0f, _size.height) - Vec2(_innerContainer->getLeftInParent(), _innerContainer->getTopInParent());
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_topBounceNeeded && _rightBounceNeeded)
         {
-            Vec2 scrollVector = Vec2(_size.width, _size.height) - Vec2(_innerContainer->getRightBoundary(), _innerContainer->getTopBoundary());
+            Vec2 scrollVector = Vec2(_size.width, _size.height) - Vec2(_innerContainer->getRightInParent(), _innerContainer->getTopInParent());
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_bottomBounceNeeded && _leftBounceNeeded)
         {
-            Vec2 scrollVector = Vec2::ZERO - Vec2(_innerContainer->getLeftBoundary(), _innerContainer->getBottomBoundary());
+            Vec2 scrollVector = Vec2::ZERO - Vec2(_innerContainer->getLeftInParent(), _innerContainer->getBottomInParent());
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_bottomBounceNeeded && _rightBounceNeeded)
         {
-            Vec2 scrollVector = Vec2(_size.width, 0.0f) - Vec2(_innerContainer->getRightBoundary(), _innerContainer->getBottomBoundary());
+            Vec2 scrollVector = Vec2(_size.width, 0.0f) - Vec2(_innerContainer->getRightInParent(), _innerContainer->getBottomInParent());
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_topBounceNeeded)
         {
-            Vec2 scrollVector = Vec2(0.0f, _size.height) - Vec2(0.0f, _innerContainer->getTopBoundary());
+            Vec2 scrollVector = Vec2(0.0f, _size.height) - Vec2(0.0f, _innerContainer->getTopInParent());
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_bottomBounceNeeded)
         {
-            Vec2 scrollVector = Vec2::ZERO - Vec2(0.0f, _innerContainer->getBottomBoundary());
+            Vec2 scrollVector = Vec2::ZERO - Vec2(0.0f, _innerContainer->getBottomInParent());
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_leftBounceNeeded)
         {
-            Vec2 scrollVector = Vec2::ZERO - Vec2(_innerContainer->getLeftBoundary(), 0.0f);
+            Vec2 scrollVector = Vec2::ZERO - Vec2(_innerContainer->getLeftInParent(), 0.0f);
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
         }
         else if (_rightBounceNeeded)
         {
-            Vec2 scrollVector = Vec2(_size.width, 0.0f) - Vec2(_innerContainer->getRightBoundary(), 0.0f);
+            Vec2 scrollVector = Vec2(_size.width, 0.0f) - Vec2(_innerContainer->getRightInParent(), 0.0f);
             float orSpeed = scrollVector.getLength()/(0.2f);
             _bounceDir = scrollVector.getNormalized();
             startBounceChildren(orSpeed);
@@ -451,7 +451,7 @@ bool ScrollView::checkNeedBounce()
 
 void ScrollView::checkBounceBoundary()
 {
-    float icBottomPos = _innerContainer->getBottomBoundary();
+    float icBottomPos = _innerContainer->getBottomInParent();
     if (icBottomPos > _bottomBoundary)
     {
         scrollToBottomEvent();
@@ -461,7 +461,7 @@ void ScrollView::checkBounceBoundary()
     {
         _bottomBounceNeeded = false;
     }
-    float icTopPos = _innerContainer->getTopBoundary();
+    float icTopPos = _innerContainer->getTopInParent();
     if (icTopPos < _topBoundary)
     {
         scrollToTopEvent();
@@ -471,7 +471,7 @@ void ScrollView::checkBounceBoundary()
     {
         _topBounceNeeded = false;
     }
-    float icRightPos = _innerContainer->getRightBoundary();
+    float icRightPos = _innerContainer->getRightInParent();
     if (icRightPos < _rightBoundary)
     {
         scrollToRightEvent();
@@ -481,7 +481,7 @@ void ScrollView::checkBounceBoundary()
     {
         _rightBounceNeeded = false;
     }
-    float icLeftPos = _innerContainer->getLeftBoundary();
+    float icLeftPos = _innerContainer->getLeftInParent();
     if (icLeftPos > _leftBoundary)
     {
         scrollToLeftEvent();
@@ -588,14 +588,14 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     {
         float realOffsetX = touchOffsetX;
         float realOffsetY = touchOffsetY;
-        float icRightPos = _innerContainer->getRightBoundary();
+        float icRightPos = _innerContainer->getRightInParent();
         if (icRightPos + realOffsetX >= _rightBoundary)
         {
             realOffsetX = _rightBoundary - icRightPos;
             bounceRightEvent();
             scrollenabled = false;
         }
-        float icTopPos = _innerContainer->getTopBoundary();
+        float icTopPos = _innerContainer->getTopInParent();
         if (icTopPos + touchOffsetY >= _topBoundary)
         {
             realOffsetY = _topBoundary - icTopPos;
@@ -608,14 +608,14 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     {
         float realOffsetX = touchOffsetX;
         float realOffsetY = touchOffsetY;
-        float icLefrPos = _innerContainer->getLeftBoundary();
+        float icLefrPos = _innerContainer->getLeftInParent();
         if (icLefrPos + realOffsetX <= _leftBoundary)
         {
             realOffsetX = _leftBoundary - icLefrPos;
             bounceLeftEvent();
             scrollenabled = false;
         }
-        float icTopPos = _innerContainer->getTopBoundary();
+        float icTopPos = _innerContainer->getTopInParent();
         if (icTopPos + touchOffsetY >= _topBoundary)
         {
             realOffsetY = _topBoundary - icTopPos;
@@ -628,14 +628,14 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     {
         float realOffsetX = touchOffsetX;
         float realOffsetY = touchOffsetY;
-        float icLefrPos = _innerContainer->getLeftBoundary();
+        float icLefrPos = _innerContainer->getLeftInParent();
         if (icLefrPos + realOffsetX <= _leftBoundary)
         {
             realOffsetX = _leftBoundary - icLefrPos;
             bounceLeftEvent();
             scrollenabled = false;
         }
-        float icBottomPos = _innerContainer->getBottomBoundary();
+        float icBottomPos = _innerContainer->getBottomInParent();
         if (icBottomPos + touchOffsetY <= _bottomBoundary)
         {
             realOffsetY = _bottomBoundary - icBottomPos;
@@ -648,14 +648,14 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     {
         float realOffsetX = touchOffsetX;
         float realOffsetY = touchOffsetY;
-        float icRightPos = _innerContainer->getRightBoundary();
+        float icRightPos = _innerContainer->getRightInParent();
         if (icRightPos + realOffsetX >= _rightBoundary)
         {
             realOffsetX = _rightBoundary - icRightPos;
             bounceRightEvent();
             scrollenabled = false;
         }
-        float icBottomPos = _innerContainer->getBottomBoundary();
+        float icBottomPos = _innerContainer->getBottomInParent();
         if (icBottomPos + touchOffsetY <= _bottomBoundary)
         {
             realOffsetY = _bottomBoundary - icBottomPos;
@@ -667,7 +667,7 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     else if (touchOffsetX == 0.0f && touchOffsetY > 0.0f) // bounce to top
     {
         float realOffsetY = touchOffsetY;
-        float icTopPos = _innerContainer->getTopBoundary();
+        float icTopPos = _innerContainer->getTopInParent();
         if (icTopPos + touchOffsetY >= _topBoundary)
         {
             realOffsetY = _topBoundary - icTopPos;
@@ -679,7 +679,7 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     else if (touchOffsetX == 0.0f && touchOffsetY < 0.0f) //bounce to bottom
     {
         float realOffsetY = touchOffsetY;
-        float icBottomPos = _innerContainer->getBottomBoundary();
+        float icBottomPos = _innerContainer->getBottomInParent();
         if (icBottomPos + touchOffsetY <= _bottomBoundary)
         {
             realOffsetY = _bottomBoundary - icBottomPos;
@@ -691,7 +691,7 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     else if (touchOffsetX > 0.0f && touchOffsetY == 0.0f) //bounce to right
     {
         float realOffsetX = touchOffsetX;
-        float icRightPos = _innerContainer->getRightBoundary();
+        float icRightPos = _innerContainer->getRightInParent();
         if (icRightPos + realOffsetX >= _rightBoundary)
         {
             realOffsetX = _rightBoundary - icRightPos;
@@ -703,7 +703,7 @@ bool ScrollView::bounceScrollChildren(float touchOffsetX, float touchOffsetY)
     else if (touchOffsetX < 0.0f && touchOffsetY == 0.0f) //bounce to left
     {
         float realOffsetX = touchOffsetX;
-        float icLeftPos = _innerContainer->getLeftBoundary();
+        float icLeftPos = _innerContainer->getLeftInParent();
         if (icLeftPos + realOffsetX <= _leftBoundary)
         {
             realOffsetX = _leftBoundary - icLeftPos;
@@ -724,7 +724,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
         {
             if (_autoScrollDir.y > 0)
             {
-                float icBottomPos = _innerContainer->getBottomBoundary();
+                float icBottomPos = _innerContainer->getBottomInParent();
                 if (icBottomPos + *touchOffsetY >= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icBottomPos;
@@ -733,7 +733,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else
             {
-                float icBottomPos = _innerContainer->getBottomBoundary();
+                float icBottomPos = _innerContainer->getBottomInParent();
                 if (icBottomPos + *touchOffsetY <= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icBottomPos;
@@ -746,7 +746,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
         {
             if (_autoScrollDir.x > 0)
             {
-                float icLeftPos = _innerContainer->getLeftBoundary();
+                float icLeftPos = _innerContainer->getLeftInParent();
                 if (icLeftPos + *touchOffsetX >= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icLeftPos;
@@ -755,7 +755,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else
             {
-                float icLeftPos = _innerContainer->getLeftBoundary();
+                float icLeftPos = _innerContainer->getLeftInParent();
                 if (icLeftPos + *touchOffsetX <= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icLeftPos;
@@ -768,13 +768,13 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
         {
             if (*touchOffsetX > 0.0f && *touchOffsetY > 0.0f) // up right
             {
-                float icLeftPos = _innerContainer->getLeftBoundary();
+                float icLeftPos = _innerContainer->getLeftInParent();
                 if (icLeftPos + *touchOffsetX >= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icLeftPos;
                     scrollenabled = false;
                 }
-                float icBottomPos = _innerContainer->getBottomBoundary();
+                float icBottomPos = _innerContainer->getBottomInParent();
                 if (icBottomPos + *touchOffsetY >= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icBottomPos;
@@ -783,13 +783,13 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX < 0.0f && *touchOffsetY > 0.0f) // up left
             {
-                float icRightPos = _innerContainer->getRightBoundary();
+                float icRightPos = _innerContainer->getRightInParent();
                 if (icRightPos + *touchOffsetX <= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icRightPos;
                     scrollenabled = false;
                 }
-                float icBottomPos = _innerContainer->getBottomBoundary();
+                float icBottomPos = _innerContainer->getBottomInParent();
                 if (icBottomPos + *touchOffsetY >= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icBottomPos;
@@ -798,13 +798,13 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX < 0.0f && *touchOffsetY < 0.0f) // down left
             {
-                float icRightPos = _innerContainer->getRightBoundary();
+                float icRightPos = _innerContainer->getRightInParent();
                 if (icRightPos + *touchOffsetX <= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icRightPos;
                     scrollenabled = false;
                 }
-                float icTopPos = _innerContainer->getTopBoundary();
+                float icTopPos = _innerContainer->getTopInParent();
                 if (icTopPos + *touchOffsetY <= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icTopPos;
@@ -813,13 +813,13 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX > 0.0f && *touchOffsetY < 0.0f) // down right
             {
-                float icLeftPos = _innerContainer->getLeftBoundary();
+                float icLeftPos = _innerContainer->getLeftInParent();
                 if (icLeftPos + *touchOffsetX >= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icLeftPos;
                     scrollenabled = false;
                 }
-                float icTopPos = _innerContainer->getTopBoundary();
+                float icTopPos = _innerContainer->getTopInParent();
                 if (icTopPos + *touchOffsetY <= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icTopPos;
@@ -828,7 +828,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX == 0.0f && *touchOffsetY > 0.0f) // up
             {
-                float icBottomPos = _innerContainer->getBottomBoundary();
+                float icBottomPos = _innerContainer->getBottomInParent();
                 if (icBottomPos + *touchOffsetY >= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icBottomPos;
@@ -837,7 +837,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX < 0.0f && *touchOffsetY == 0.0f) // left
             {
-                float icRightPos = _innerContainer->getRightBoundary();
+                float icRightPos = _innerContainer->getRightInParent();
                 if (icRightPos + *touchOffsetX <= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icRightPos;
@@ -846,7 +846,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX == 0.0f && *touchOffsetY < 0.0f) // down
             {
-                float icTopPos = _innerContainer->getTopBoundary();
+                float icTopPos = _innerContainer->getTopInParent();
                 if (icTopPos + *touchOffsetY <= _autoScrollDestination.y)
                 {
                     *touchOffsetY = _autoScrollDestination.y - icTopPos;
@@ -855,7 +855,7 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
             }
             else if (*touchOffsetX > 0.0f && *touchOffsetY == 0.0f) // right
             {
-                float icLeftPos = _innerContainer->getLeftBoundary();
+                float icLeftPos = _innerContainer->getLeftInParent();
                 if (icLeftPos + *touchOffsetX >= _autoScrollDestination.x)
                 {
                     *touchOffsetX = _autoScrollDestination.x - icLeftPos;
@@ -869,346 +869,320 @@ bool ScrollView::checkCustomScrollDestination(float* touchOffsetX, float* touchO
     }
     return scrollenabled;
 }
-    
-bool ScrollView::scrollChilderVertical(float touchOffsetX, float touchOffsetY)
-{
-    bool scrollenabled = true;
-
-    float realOffset = touchOffsetY;
-    if (_bounceEnabled)
-    {
-        float icBottomPos = _innerContainer->getBottomBoundary();
-        if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
-        {
-            realOffset = _bounceBottomBoundary - icBottomPos;
-            scrollToBottomEvent();
-            scrollenabled = false;
-        }
-        float icTopPos = _innerContainer->getTopBoundary();
-        if (icTopPos + touchOffsetY <= _bounceTopBoundary)
-        {
-            realOffset = _bounceTopBoundary - icTopPos;
-            scrollToTopEvent();
-            scrollenabled = false;
-        }
-    }
-    else
-    {
-        float icBottomPos = _innerContainer->getBottomBoundary();
-        if (icBottomPos + touchOffsetY >= _bottomBoundary)
-        {
-            realOffset = _bottomBoundary - icBottomPos;
-            scrollToBottomEvent();
-            scrollenabled = false;
-        }
-        float icTopPos = _innerContainer->getTopBoundary();
-        if (icTopPos + touchOffsetY <= _topBoundary)
-        {
-            realOffset = _topBoundary - icTopPos;
-            scrollToTopEvent();
-            scrollenabled = false;
-        }
-    }
-    moveChildren(0.0f, realOffset);
-    
-    return scrollenabled;
-}
-    
-bool ScrollView::scrollChilderHorizontal(float touchOffsetX, float touchOffsetY)
-{
-    bool scrollenabled = true;
-
-    float realOffset = touchOffsetX;
-    if (_bounceEnabled)
-    {
-        float icRightPos = _innerContainer->getRightBoundary();
-        if (icRightPos + touchOffsetX <= _bounceRightBoundary)
-        {
-            realOffset = _bounceRightBoundary - icRightPos;
-            scrollToRightEvent();
-            scrollenabled = false;
-        }
-        float icLeftPos = _innerContainer->getLeftBoundary();
-        if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
-        {
-            realOffset = _bounceLeftBoundary - icLeftPos;
-            scrollToLeftEvent();
-            scrollenabled = false;
-        }
-    }
-    else
-    {
-        float icRightPos = _innerContainer->getRightBoundary();
-        if (icRightPos + touchOffsetX <= _rightBoundary)
-        {
-            realOffset = _rightBoundary - icRightPos;
-            scrollToRightEvent();
-            scrollenabled = false;
-        }
-        float icLeftPos = _innerContainer->getLeftBoundary();
-        if (icLeftPos + touchOffsetX >= _leftBoundary)
-        {
-            realOffset = _leftBoundary - icLeftPos;
-            scrollToLeftEvent();
-            scrollenabled = false;
-        }
-    }
-    moveChildren(realOffset, 0.0f);
-    
-    return scrollenabled;
-}
-    
-bool ScrollView::scrollChilderBoth(float touchOffsetX, float touchOffsetY)
-{
-    bool scrollenabled = true;
-    float realOffsetX = touchOffsetX;
-    float realOffsetY = touchOffsetY;
-    if (_bounceEnabled)
-    {
-        if (touchOffsetX > 0.0f && touchOffsetY > 0.0f) // up right
-        {
-            float icLeftPos = _innerContainer->getLeftBoundary();
-            if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
-            {
-                realOffsetX = _bounceLeftBoundary - icLeftPos;
-                scrollToLeftEvent();
-                scrollenabled = false;
-            }
-            float icBottomPos = _innerContainer->getBottomBoundary();
-            if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
-            {
-                realOffsetY = _bounceBottomBoundary - icBottomPos;
-                scrollToBottomEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX < 0.0f && touchOffsetY > 0.0f) // up left
-        {
-            float icRightPos = _innerContainer->getRightBoundary();
-            if (icRightPos + touchOffsetX <= _bounceRightBoundary)
-            {
-                realOffsetX = _bounceRightBoundary - icRightPos;
-                scrollToRightEvent();
-                scrollenabled = false;
-            }
-            float icBottomPos = _innerContainer->getBottomBoundary();
-            if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
-            {
-                realOffsetY = _bounceBottomBoundary - icBottomPos;
-                scrollToBottomEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX < 0.0f && touchOffsetY < 0.0f) // down left
-        {
-            float icRightPos = _innerContainer->getRightBoundary();
-            if (icRightPos + touchOffsetX <= _bounceRightBoundary)
-            {
-                realOffsetX = _bounceRightBoundary - icRightPos;
-                scrollToRightEvent();
-                scrollenabled = false;
-            }
-            float icTopPos = _innerContainer->getTopBoundary();
-            if (icTopPos + touchOffsetY <= _bounceTopBoundary)
-            {
-                realOffsetY = _bounceTopBoundary - icTopPos;
-                scrollToTopEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX > 0.0f && touchOffsetY < 0.0f) // down right
-        {
-            float icLeftPos = _innerContainer->getLeftBoundary();
-            if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
-            {
-                realOffsetX = _bounceLeftBoundary - icLeftPos;
-                scrollToLeftEvent();
-                scrollenabled = false;
-            }
-            float icTopPos = _innerContainer->getTopBoundary();
-            if (icTopPos + touchOffsetY <= _bounceTopBoundary)
-            {
-                realOffsetY = _bounceTopBoundary - icTopPos;
-                scrollToTopEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX == 0.0f && touchOffsetY > 0.0f) // up
-        {
-            float icBottomPos = _innerContainer->getBottomBoundary();
-            if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
-            {
-                realOffsetY = _bounceBottomBoundary - icBottomPos;
-                scrollToBottomEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX < 0.0f && touchOffsetY == 0.0f) // left
-        {
-            float icRightPos = _innerContainer->getRightBoundary();
-            if (icRightPos + touchOffsetX <= _bounceRightBoundary)
-            {
-                realOffsetX = _bounceRightBoundary - icRightPos;
-                scrollToRightEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX == 0.0f && touchOffsetY < 0.0f) // down
-        {
-            float icTopPos = _innerContainer->getTopBoundary();
-            if (icTopPos + touchOffsetY <= _bounceTopBoundary)
-            {
-                realOffsetY = _bounceTopBoundary - icTopPos;
-                scrollToTopEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX > 0.0f && touchOffsetY == 0.0f) // right
-        {
-            float icLeftPos = _innerContainer->getLeftBoundary();
-            if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
-            {
-                realOffsetX = _bounceLeftBoundary - icLeftPos;
-                scrollToLeftEvent();
-                scrollenabled = false;
-            }
-        }
-    }
-    else
-    {
-        if (touchOffsetX > 0.0f && touchOffsetY > 0.0f) // up right
-        {
-            float icLeftPos = _innerContainer->getLeftBoundary();
-            if (icLeftPos + touchOffsetX >= _leftBoundary)
-            {
-                realOffsetX = _leftBoundary - icLeftPos;
-                scrollToLeftEvent();
-                scrollenabled = false;
-            }
-            float icBottomPos = _innerContainer->getBottomBoundary();
-            if (icBottomPos + touchOffsetY >= _bottomBoundary)
-            {
-                realOffsetY = _bottomBoundary - icBottomPos;
-                scrollToBottomEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX < 0.0f && touchOffsetY > 0.0f) // up left
-        {
-            float icRightPos = _innerContainer->getRightBoundary();
-            if (icRightPos + touchOffsetX <= _rightBoundary)
-            {
-                realOffsetX = _rightBoundary - icRightPos;
-                scrollToRightEvent();
-                scrollenabled = false;
-            }
-            float icBottomPos = _innerContainer->getBottomBoundary();
-            if (icBottomPos + touchOffsetY >= _bottomBoundary)
-            {
-                realOffsetY = _bottomBoundary - icBottomPos;
-                scrollToBottomEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX < 0.0f && touchOffsetY < 0.0f) // down left
-        {
-            float icRightPos = _innerContainer->getRightBoundary();
-            if (icRightPos + touchOffsetX <= _rightBoundary)
-            {
-                realOffsetX = _rightBoundary - icRightPos;
-                scrollToRightEvent();
-                scrollenabled = false;
-            }
-            float icTopPos = _innerContainer->getTopBoundary();
-            if (icTopPos + touchOffsetY <= _topBoundary)
-            {
-                realOffsetY = _topBoundary - icTopPos;
-                scrollToTopEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX > 0.0f && touchOffsetY < 0.0f) // down right
-        {
-            float icLeftPos = _innerContainer->getLeftBoundary();
-            if (icLeftPos + touchOffsetX >= _leftBoundary)
-            {
-                realOffsetX = _leftBoundary - icLeftPos;
-                scrollToLeftEvent();
-                scrollenabled = false;
-            }
-            float icTopPos = _innerContainer->getTopBoundary();
-            if (icTopPos + touchOffsetY <= _topBoundary)
-            {
-                realOffsetY = _topBoundary - icTopPos;
-                scrollToTopEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX == 0.0f && touchOffsetY > 0.0f) // up
-        {
-            float icBottomPos = _innerContainer->getBottomBoundary();
-            if (icBottomPos + touchOffsetY >= _bottomBoundary)
-            {
-                realOffsetY = _bottomBoundary - icBottomPos;
-                scrollToBottomEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX < 0.0f && touchOffsetY == 0.0f) // left
-        {
-            float icRightPos = _innerContainer->getRightBoundary();
-            if (icRightPos + touchOffsetX <= _rightBoundary)
-            {
-                realOffsetX = _rightBoundary - icRightPos;
-                scrollToRightEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX == 0.0f && touchOffsetY < 0.0f) // down
-        {
-            float icTopPos = _innerContainer->getTopBoundary();
-            if (icTopPos + touchOffsetY <= _topBoundary)
-            {
-                realOffsetY = _topBoundary - icTopPos;
-                scrollToTopEvent();
-                scrollenabled = false;
-            }
-        }
-        else if (touchOffsetX > 0.0f && touchOffsetY == 0.0f) // right
-        {
-            float icLeftPos = _innerContainer->getLeftBoundary();
-            if (icLeftPos + touchOffsetX >= _leftBoundary)
-            {
-                realOffsetX = _leftBoundary - icLeftPos;
-                scrollToLeftEvent();
-                scrollenabled = false;
-            }
-        }
-    }
-    moveChildren(realOffsetX, realOffsetY);
-    return scrollenabled;
-}
 
 bool ScrollView::scrollChildren(float touchOffsetX, float touchOffsetY)
 {
     bool scrollenabled = true;
     scrollingEvent();
-    
     switch (_direction)
     {
         case Direction::VERTICAL: // vertical
         {
-            scrollenabled = this->scrollChilderVertical(touchOffsetX, touchOffsetY);
+            float realOffset = touchOffsetY;
+            if (_bounceEnabled)
+            {
+                float icBottomPos = _innerContainer->getBottomInParent();
+                if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
+                {
+                    realOffset = _bounceBottomBoundary - icBottomPos;
+                    scrollToBottomEvent();
+                    scrollenabled = false;
+                }
+                float icTopPos = _innerContainer->getTopInParent();
+                if (icTopPos + touchOffsetY <= _bounceTopBoundary)
+                {
+                    realOffset = _bounceTopBoundary - icTopPos;
+                    scrollToTopEvent();
+                    scrollenabled = false;
+                }
+            }
+            else
+            {
+                float icBottomPos = _innerContainer->getBottomInParent();
+                if (icBottomPos + touchOffsetY >= _bottomBoundary)
+                {
+                    realOffset = _bottomBoundary - icBottomPos;
+                    scrollToBottomEvent();
+                    scrollenabled = false;
+                }
+                float icTopPos = _innerContainer->getTopInParent();
+                if (icTopPos + touchOffsetY <= _topBoundary)
+                {
+                    realOffset = _topBoundary - icTopPos;
+                    scrollToTopEvent();
+                    scrollenabled = false;
+                }
+            }
+            moveChildren(0.0f, realOffset);
             break;
         }
         case Direction::HORIZONTAL: // horizontal
         {
-            scrollenabled = this->scrollChilderHorizontal(touchOffsetX, touchOffsetY);
+            float realOffset = touchOffsetX;
+            if (_bounceEnabled)
+            {
+                float icRightPos = _innerContainer->getRightInParent();
+                if (icRightPos + touchOffsetX <= _bounceRightBoundary)
+                {
+                    realOffset = _bounceRightBoundary - icRightPos;
+                    scrollToRightEvent();
+                    scrollenabled = false;
+                }
+                float icLeftPos = _innerContainer->getLeftInParent();
+                if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
+                {
+                    realOffset = _bounceLeftBoundary - icLeftPos;
+                    scrollToLeftEvent();
+                    scrollenabled = false;
+                }
+            }
+            else
+            {
+                float icRightPos = _innerContainer->getRightInParent();
+                if (icRightPos + touchOffsetX <= _rightBoundary)
+                {
+                    realOffset = _rightBoundary - icRightPos;
+                    scrollToRightEvent();
+                    scrollenabled = false;
+                }
+                float icLeftPos = _innerContainer->getLeftInParent();
+                if (icLeftPos + touchOffsetX >= _leftBoundary)
+                {
+                    realOffset = _leftBoundary - icLeftPos;
+                    scrollToLeftEvent();
+                    scrollenabled = false;
+                }
+            }
+            moveChildren(realOffset, 0.0f);
             break;
         }
         case Direction::BOTH:
         {
-            scrollenabled = this->scrollChilderBoth(touchOffsetX, touchOffsetY);
+            float realOffsetX = touchOffsetX;
+            float realOffsetY = touchOffsetY;
+            if (_bounceEnabled)
+            {
+                if (touchOffsetX > 0.0f && touchOffsetY > 0.0f) // up right
+                {
+                    float icLeftPos = _innerContainer->getLeftInParent();
+                    if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
+                    {
+                        realOffsetX = _bounceLeftBoundary - icLeftPos;
+                        scrollToLeftEvent();
+                        scrollenabled = false;
+                    }
+                    float icBottomPos = _innerContainer->getBottomInParent();
+                    if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
+                    {
+                        realOffsetY = _bounceBottomBoundary - icBottomPos;
+                        scrollToBottomEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX < 0.0f && touchOffsetY > 0.0f) // up left
+                {
+                    float icRightPos = _innerContainer->getRightInParent();
+                    if (icRightPos + touchOffsetX <= _bounceRightBoundary)
+                    {
+                        realOffsetX = _bounceRightBoundary - icRightPos;
+                        scrollToRightEvent();
+                        scrollenabled = false;
+                    }
+                    float icBottomPos = _innerContainer->getBottomInParent();
+                    if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
+                    {
+                        realOffsetY = _bounceBottomBoundary - icBottomPos;
+                        scrollToBottomEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX < 0.0f && touchOffsetY < 0.0f) // down left
+                {
+                    float icRightPos = _innerContainer->getRightInParent();
+                    if (icRightPos + touchOffsetX <= _bounceRightBoundary)
+                    {
+                        realOffsetX = _bounceRightBoundary - icRightPos;
+                        scrollToRightEvent();
+                        scrollenabled = false;
+                    }
+                    float icTopPos = _innerContainer->getTopInParent();
+                    if (icTopPos + touchOffsetY <= _bounceTopBoundary)
+                    {
+                        realOffsetY = _bounceTopBoundary - icTopPos;
+                        scrollToTopEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX > 0.0f && touchOffsetY < 0.0f) // down right
+                {
+                    float icLeftPos = _innerContainer->getLeftInParent();
+                    if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
+                    {
+                        realOffsetX = _bounceLeftBoundary - icLeftPos;
+                        scrollToLeftEvent();
+                        scrollenabled = false;
+                    }
+                    float icTopPos = _innerContainer->getTopInParent();
+                    if (icTopPos + touchOffsetY <= _bounceTopBoundary)
+                    {
+                        realOffsetY = _bounceTopBoundary - icTopPos;
+                        scrollToTopEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX == 0.0f && touchOffsetY > 0.0f) // up
+                {
+                    float icBottomPos = _innerContainer->getBottomInParent();
+                    if (icBottomPos + touchOffsetY >= _bounceBottomBoundary)
+                    {
+                        realOffsetY = _bounceBottomBoundary - icBottomPos;
+                        scrollToBottomEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX < 0.0f && touchOffsetY == 0.0f) // left
+                {
+                    float icRightPos = _innerContainer->getRightInParent();
+                    if (icRightPos + touchOffsetX <= _bounceRightBoundary)
+                    {
+                        realOffsetX = _bounceRightBoundary - icRightPos;
+                        scrollToRightEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX == 0.0f && touchOffsetY < 0.0f) // down
+                {
+                    float icTopPos = _innerContainer->getTopInParent();
+                    if (icTopPos + touchOffsetY <= _bounceTopBoundary)
+                    {
+                        realOffsetY = _bounceTopBoundary - icTopPos;
+                        scrollToTopEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX > 0.0f && touchOffsetY == 0.0f) // right
+                {
+                    float icLeftPos = _innerContainer->getLeftInParent();
+                    if (icLeftPos + touchOffsetX >= _bounceLeftBoundary)
+                    {
+                        realOffsetX = _bounceLeftBoundary - icLeftPos;
+                        scrollToLeftEvent();
+                        scrollenabled = false;
+                    }
+                }
+            }
+            else
+            {
+                if (touchOffsetX > 0.0f && touchOffsetY > 0.0f) // up right
+                {
+                    float icLeftPos = _innerContainer->getLeftInParent();
+                    if (icLeftPos + touchOffsetX >= _leftBoundary)
+                    {
+                        realOffsetX = _leftBoundary - icLeftPos;
+                        scrollToLeftEvent();
+                        scrollenabled = false;
+                    }
+                    float icBottomPos = _innerContainer->getBottomInParent();
+                    if (icBottomPos + touchOffsetY >= _bottomBoundary)
+                    {
+                        realOffsetY = _bottomBoundary - icBottomPos;
+                        scrollToBottomEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX < 0.0f && touchOffsetY > 0.0f) // up left
+                {
+                    float icRightPos = _innerContainer->getRightInParent();
+                    if (icRightPos + touchOffsetX <= _rightBoundary)
+                    {
+                        realOffsetX = _rightBoundary - icRightPos;
+                        scrollToRightEvent();
+                        scrollenabled = false;
+                    }
+                    float icBottomPos = _innerContainer->getBottomInParent();
+                    if (icBottomPos + touchOffsetY >= _bottomBoundary)
+                    {
+                        realOffsetY = _bottomBoundary - icBottomPos;
+                        scrollToBottomEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX < 0.0f && touchOffsetY < 0.0f) // down left
+                {
+                    float icRightPos = _innerContainer->getRightInParent();
+                    if (icRightPos + touchOffsetX <= _rightBoundary)
+                    {
+                        realOffsetX = _rightBoundary - icRightPos;
+                        scrollToRightEvent();
+                        scrollenabled = false;
+                    }
+                    float icTopPos = _innerContainer->getTopInParent();
+                    if (icTopPos + touchOffsetY <= _topBoundary)
+                    {
+                        realOffsetY = _topBoundary - icTopPos;
+                        scrollToTopEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX > 0.0f && touchOffsetY < 0.0f) // down right
+                {
+                    float icLeftPos = _innerContainer->getLeftInParent();
+                    if (icLeftPos + touchOffsetX >= _leftBoundary)
+                    {
+                        realOffsetX = _leftBoundary - icLeftPos;
+                        scrollToLeftEvent();
+                        scrollenabled = false;
+                    }
+                    float icTopPos = _innerContainer->getTopInParent();
+                    if (icTopPos + touchOffsetY <= _topBoundary)
+                    {
+                        realOffsetY = _topBoundary - icTopPos;
+                        scrollToTopEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX == 0.0f && touchOffsetY > 0.0f) // up
+                {
+                    float icBottomPos = _innerContainer->getBottomInParent();
+                    if (icBottomPos + touchOffsetY >= _bottomBoundary)
+                    {
+                        realOffsetY = _bottomBoundary - icBottomPos;
+                        scrollToBottomEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX < 0.0f && touchOffsetY == 0.0f) // left
+                {
+                    float icRightPos = _innerContainer->getRightInParent();
+                    if (icRightPos + touchOffsetX <= _rightBoundary)
+                    {
+                        realOffsetX = _rightBoundary - icRightPos;
+                        scrollToRightEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX == 0.0f && touchOffsetY < 0.0f) // down
+                {
+                    float icTopPos = _innerContainer->getTopInParent();
+                    if (icTopPos + touchOffsetY <= _topBoundary)
+                    {
+                        realOffsetY = _topBoundary - icTopPos;
+                        scrollToTopEvent();
+                        scrollenabled = false;
+                    }
+                }
+                else if (touchOffsetX > 0.0f && touchOffsetY == 0.0f) // right
+                {
+                    float icLeftPos = _innerContainer->getLeftInParent();
+                    if (icLeftPos + touchOffsetX >= _leftBoundary)
+                    {
+                        realOffsetX = _leftBoundary - icLeftPos;
+                        scrollToLeftEvent();
+                        scrollenabled = false;
+                    }
+                }
+            }
+            moveChildren(realOffsetX, realOffsetY);
             break;
         }
         default:
