+++ cocos/ui/UIScrollView.h
@@ -357,13 +357,7 @@ protected:
     void startBounceChildren(float v);
     void stopBounceChildren();
     bool checkCustomScrollDestination(float* touchOffsetX, float* touchOffsetY);
-    
     virtual bool scrollChildren(float touchOffsetX, float touchOffsetY);
-    virtual bool scrollChilderHorizontal(float touchOffsetX, float touchOffsetY);
-    virtual bool scrollChilderVertical(float touchOffsetX, float touchOffsetY);
-    virtual bool scrollChilderBoth(float touchOffsetX, float touchOffsetY);
-
-    
     bool bounceScrollChildren(float touchOffsetX, float touchOffsetY);
     void startRecordSlidAction();
     virtual void endRecordSlidAction();
