+++ cocos/ui/proj.win32/libGUI.vcxproj.filters
@@ -87,9 +87,6 @@
     <ClInclude Include="..\UIVBox.h">
       <Filter>Layouts</Filter>
     </ClInclude>
-    <ClInclude Include="..\UILayoutExecutant.h">
-      <Filter>Layouts</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\UIScrollView.cpp">
@@ -156,9 +153,6 @@
       <Filter>Layouts</Filter>
     </ClCompile>
     <ClCompile Include="..\UIVBox.cpp">
-      <Filter>Layouts</Filter>
-    </ClCompile>
-    <ClCompile Include="..\UILayoutExecutant.cpp">
       <Filter>Layouts</Filter>
     </ClCompile>
   </ItemGroup>
